+++ src/AutoConstructor.Generator/Source.cs
@@ -36,7 +36,7 @@ public static class Source
             /// <param name="accessibility">Configure the accessibility of the constructor, public by default</param>
             /// <param name="addDefaultBaseAttribute">Configure automatic injection of <c>AutoConstructorDefaultBaseAttribute</c></param>
             /// <param name="disableThisCall">Disable call to this when it would have been called</param>
-            /// <param name="addParameterless">Generate parameterless constructor for serialization purposes too.</param>
+            /// <param name="addParameterless">Generate a parameterless constructor too.</param>
             public {{AttributeFullName}}(string accessibility = null, bool addDefaultBaseAttribute = false, bool disableThisCall = false, bool addParameterless = false)
             {
                 Accessibility = accessibility;